+++ src/com/guardian/history/records/BAlarmCountHistoryRecord.java
@@ -19,11 +19,6 @@ import java.io.IOException;
  */
 @NiagaraType
 @NiagaraProperty(
-        name = "typeTag",
-        type = "BHistoryTagEnum",
-        defaultValue = "BHistoryTagEnum.DEFAULT"
-)
-@NiagaraProperty(
         name = "value",
         type = "boolean",
         defaultValue = "false"
@@ -49,35 +44,11 @@ import java.io.IOException;
         defaultValue = "BRelTime.DEFAULT"
 )
 public final class BAlarmCountHistoryRecord
-        extends BTrendRecord
-        implements JSONSupport
+        extends BValueHistoryRecord
 {
 /*+ ------------ BEGIN BAJA AUTO GENERATED CODE ------------ +*/
-/*@ $BAlarmCountHistoryRecord(588473534)1.0$ @*/
-/* Generated Thu Nov 09 09:51:36 CST 2017 by Slot-o-Matic (c) Tridium, Inc. 2012 */
-
-////////////////////////////////////////////////////////////////
-// Property "typeTag"
-////////////////////////////////////////////////////////////////
-  
-  /**
-   * Slot for the {@code typeTag} property.
-   * @see #getTypeTag
-   * @see #setTypeTag
-   */
-  public static final Property typeTag = newProperty(0, BHistoryTagEnum.DEFAULT, null);
-  
-  /**
-   * Get the {@code typeTag} property.
-   * @see #typeTag
-   */
-  public BHistoryTagEnum getTypeTag() { return (BHistoryTagEnum)get(typeTag); }
-  
-  /**
-   * Set the {@code typeTag} property.
-   * @see #typeTag
-   */
-  public void setTypeTag(BHistoryTagEnum v) { set(typeTag, v, null); }
+/*@ $com.guardian.history.records.BAlarmCountHistoryRecord(1030788856)1.0$ @*/
+/* Generated Mon Dec 04 10:21:31 CST 2017 by Slot-o-Matic (c) Tridium, Inc. 2012 */
 
 ////////////////////////////////////////////////////////////////
 // Property "value"